Update README.md
Additions:
- Added MAP03 "Blue-Line Beatdown"

Changes / Fixes:
* Fixed soft-lock in MAP04's fountain
- Moved "Civilian Center" to the MAP04 slot
- Changed MAP01's song from "Underworld" to "Rage"
- Changed MAP02's song from "Gun Pointed to the Head of the Universe" to "Clay"
- Changed MAP04's song from "Goin' Down the Fast Way" to "Civilian Center"
- Changed the Title song from "Fiddler on the Green" to "Catnip Forest"
- Changed the Intermission song from "Out on the Water" to "Lunar Transit"

Why did I change all the songs? Many of the songs were just midi renditions of non-midi songs. These new songs (which the exception of Rage) were made specifically for the midi format, and are as a result higher quality. A few songs were also the work of bands whose labels are known to abuse DMCA (despite the fact this is a non-commercial project and would fall under fair-use), so to be safe I changed them to the works of community members who explicitly allow their works to be used.